P01: Set up Description
The states/sections of the program will be inserted in blocks. Once the single section (eg "title") will be completed, the next on will start. They will not necessarily follow the inteded order, for the first ones will be the easiest/quickest ones, whereas the most challenging ones will be inserted last.